--- a/packages/playwright-core/package.json
+++ b/packages/playwright-core/package.json
@@ -40,5 +40,8 @@
   "bin": {
     "playwright-core": "cli.js"
   },
-  "types": "types/types.d.ts"
+  "types": "types/types.d.ts",
+  "dependencies": {
+    "@blazediff/core": "^0.5.0"
+  }
 }