--- a/packages/playwright-core/src/server/utils/DEPS.list
+++ b/packages/playwright-core/src/server/utils/DEPS.list
@@ -6,4 +6,3 @@
 
 [comparators.ts]
 ./image_tools
-../../third_party/pixelmatch