… packages/playwright-core/src/third_party/pixelmatch.js