# Description

Replaces `pixelmatch` third-party with a ~20-25% faster alternative [BlazeDiff](https://github.com/teimurjan/blazediff). It's type-safe and uses an enhanced algorithm inspired by pixelmatch. The API and results are identical.

Check [Run Benchmark step](https://github.com/teimurjan/blazediff/actions/runs/17368266694/job/49299242732).

# Benchmarks

I have created a patched [`playwright-core`](https://github.com/teimurjan/playwright-core-blazediff) package with [`@blazediff/core`](https://www.npmjs.com/package/@blazediff/core) instead of [pixelmatch](https://www.npmjs.com/package/pixelmatch).

Then, I created two repositories for benchmarking:
1. Using the original Playwright: https://github.com/teimurjan/playwright-benchmark.
2. Using the patched Playwright: https://github.com/teimurjan/playwright-blazediff-benchmark.

*The benchmarks track time spent on `toHaveScreenshot` with 3 warmup iterations and 25 actual iterations.*

Results:
- Before (pixelmatch): 2059.12ms, 1522.12ms, 1941.32ms average times
  - Check [Run Playwright tests step](https://github.com/teimurjan/playwright-benchmark/actions/runs/17411827788/job/49430691170).
- After (BlazeDiff): 1987.28ms, 1476.44ms, 1893.28ms average times
  - Check [Run Playwright tests step](https://github.com/teimurjan/playwright-blazediff-benchmark/actions/runs/17411895665/job/49430936730).
- **Improvement**: ~4% faster in real Playwright usage scenarios.

Closes #37267.